requirements.txt
flask==2.3.3
werkzeug==2.3.7
requests==2.31.0
google-cloud-vision==3.4.4 